#1312 Changed format to appease Maven, need to fix Tests.
