+++ row-data-gateway/etc/row-data-gateway.urm.puml
@@ -1,0 +1,36 @@
+@startuml
+package com.iluwatar.rowdatagateway {
+  class App {
+    - urlDB : String {static}
+    - App()
+    + clearAll() {static}
+    + display() {static}
+    + initialize() {static}
+    + main(args : String[]) {static}
+  }
+  class RowData {
+    - id : int
+    - name : String
+    - value : int
+    + RowData(i : int, s : String, i1 : int)
+    + getID() : int
+    + getName() : String
+    + getValue() : int
+    + setID(id : int)
+    + setName(name : String)
+    + setValue(value : int)
+  }
+  class RowDataGateway {
+    + c : Connection {static}
+    - rowdata : RowData
+    - urlDB : String
+    + RowDataGateway(rowdata : RowData, urlDB : String)
+    + delete()
+    + getRowData() : RowData
+    + insert()
+    + setRowData(rowData : RowData)
+    + update()
+  }
+}
+RowDataGateway -->  "-rowdata" RowData
+@enduml
